now build dynamodb local with logging enabled
