+++ magefiles/docker-compose.dev.yml
@@ -1,7 +1,9 @@
 services:
   dynamodb:
     command: "-jar DynamoDBLocal.jar -inMemory"
-    image: "amazon/dynamodb-local:latest"
     ports:
       - "8000:8000"
     working_dir: /home/dynamodblocal
+    build:
+      context: ./dynamodblocal
+      dockerfile: Dockerfile
